Extract shared components
