+++ src/app/cluster/details/shared/version-change-dialog/style.scss
