+++ src/app/cluster/details/shared/version-change-dialog/template.html
