+++ src/app/cluster/details/shared/version-picker/style.scss
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-@use 'src/assets/css/variables';
-@use 'src/assets/css/mixins';
+@use 'variables';
+@use 'mixins';
 
 .km-version-picker-container {
   margin: 6px 30px 6px 0;
