+++ src/app/cluster/details/shared/version-picker/template.html
