+++ src/app/cluster/details/shared/version-picker/theme.scss
