+++ src/assets/css/theme/_component.scss
@@ -15,7 +15,7 @@
 @import '../../../app/cluster/details/cluster/cni-version/theme';
 @import '../../../app/cluster/details/cluster/constraints/theme';
 @import '../../../app/cluster/details/cluster/machine-deployment-details/cluster-panel/theme';
-@import '../../../app/cluster/details/cluster/version-picker/theme';
+@import '../../../app/cluster/details/shared/version-picker/theme';
 @import 'src/app/core/components/footer/theme';
 @import 'src/app/core/components/navigation/project/theme';
 @import 'src/app/core/components/navigation/theme';
